Make Zoom One Type updatable.
Fixes #2

Not sure if this works on creation or not, but update and read work. Not
sure if create worked as it was coded. This will wipe out phone feature
on create as coded. This is what NDSU wants, as we require a different
API call before phone feature can be turned on anyway.

Also change type over to long, as our Zoom One Type is way longer than
an integer in Java.